requirements-dev.txt
fastapi==0.92.0
pydantic==1.10.5
uvicorn[standard]==0.20.0
gpt_index==0.4.18
google-cloud-storage==2.7.0
python-multipart==0.0.6
langchain==0.0.98
email-validator==1.3.1
faiss-cpu==1.7.2
PyPDF2==3.0.1
pydub==0.25.1
google-cloud-texttospeech==2.14.1
google-cloud-translate==3.11.0
google-cloud-speech==2.18.0
asyncpg==0.27.0
pytz==2022.7.1
ffmpeg
scikit-learn==1.2.1
urllib3==1.26.15